+++ release-notes.md
@@ -1,4 +1,4 @@
-Version 1.4.0 (unreleased)
+Version 1.4.0 (latest)
 --------------
 * Fixed a an issue where EntityContainer.stop() wasn't clearing the internal entity set
     reference causing other issues if operations were performed on the container after
@@ -10,7 +10,7 @@ Version 1.4.0 (unreleased)
     and a default wiring to the standard JME camera. The movement target is configurable.
 
 
-Version 1.3.0 (latest)
+Version 1.3.0
 --------------
 * Upped the Zay-ES-net version to 1.4.0 to get the entity set filtering
     bug fix.